armnn-onnx: use stable version 1.6.0 and repair build
Version previously indicated in this recipe was not the stable one (no
tag applied) and rather a snapshot of the tip from development branch.

Resolve this by using a stable tag for version 1.6.0 and wait until
1.6.1 gets it's tag applied.

In addition, indicate that protobuf3 should be used for build, exclude
-ml namespace from the build and repair installation by verifying that
only files are getting installed (no folders are taken).

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>